+++ naverCafe/src/Constants.tsx
@@ -1,2 +1,13 @@
 //baseURL 및 기타 상수
+import { type Cafe } from "./Types";
+
 export const baseURL = "";
+
+export const waffleCafe: Cafe = {
+  name: "와플카페",
+  url: "https://cafe.naver.com/wafflecafe",
+  manager: "매니저",
+  createdAt: "2024.01.13 개설",
+  location: "서울특별시 관악구",
+  description: "와플스튜디오 team7의 토이프로젝트 카페입니다.",
+};
