+++ naverCafe/src/components/body/contents/UserPage.tsx
@@ -1,0 +1,1 @@
+//UserPage에서 MyPage, MemberPage로 나뉘면 될 것 같습니다.
