+++ naverCafe/수정된 사항 및 엔티티.md
@@ -1,8 +1,13 @@
 Context API와 관련해서 찾아보니, 굳이 저번 과제에서 API와 관련된 함수를 context로 관리할 필요가 없었다는 것을 알게 되었습니다. 아래는 제가 참고한 자료입니다.
 
+
+
 - [리액트에서 Context API 사용하는 방법](https://velog.io/@velopert/react-context-tutorial)
 - [지난 과제 예시 답안 (세미나장님)](https://github.com/joongwon/waffle-react-hw.git)
 	- 여기서 contexts 폴더와 api 파일, main 파일을 특히 참고했습니다.
+
+
+
 
 그래서 일단은 src/API에 엔티티 별로 수행하는 함수를 따로 만들고, 
 전역적으로 로그인 상태를 확인하기 위해 로그인 및 회원가입 기능에서만 따로 authContext를 만드는 것이 좋아보입니다.
@@ -18,8 +23,11 @@ Types.tsx에 필요해보이는 type을 정리해두긴 했는데, 뭉치(?)가 
 아래는 구독 관계 및 태그를 제외하고 작성한 엔티티 명세(?)입니다.
 쓸모가 있을진 모르겠지만... 나중에 API가 확정되면 그에 맞춰 얻을 수 있는 정보를 매핑하기 편할 것 같아서 작성해보았습니다... 제가 맡은 부분이 아닌 곳도 있긴 하지만 생각보다 할게 너무 많은 것 같아서 이거라도 해보았습니다
 
-| ==**User**== | **signUp** | **Login** | **SideBar** | **MyPage** | **MemberPage** | **Profile(Change)** |
-| ---- | ---- | ---- | ---- | ---- | ---- | ---- |
+
+
+
+| ***User*** | **signUp** | **Login** | **SideBar** | **MyPage** | **MemberPage** | **Profile(Change)** |
+| :--: | :--: | :--: | :--: | :--: | :--: | :--: |
 | id <small>아이디<small> | O | O |  | O | O |  |
 | password <small>비밀번호<small> | O | O |  |  |  |  |
 | email <small>이메일<small> | O |  |  |  |  |  |
@@ -39,10 +47,13 @@ Types.tsx에 필요해보이는 type을 정리해두긴 했는데, 뭉치(?)가 
 | likedArticle <small>좋아요한 글<small> |  |  |  | O |  |  |
 | (deletedArticle) <small>삭제한 게시글<small> |  |  |  | O |  |  |
 | favoriteBoards <small>즐겨찾는 게시판<small> |  |  | O |  |  |  |
+
 즐겨찾는 게시판의 경우, Board에서 다루는 것 같습니다.
 
-| ==**Article**== | **게시글 페이지** | **글 목록 페이지** | **MyPage** (작성글) | **MemberPage** | **삭제한 게시글 목록** | **삭제한 게시글 페이지** |
-| ---- | ---- | ---- | ---- | ---- | ---- | ---- |
+
+
+| ***Article*** | **게시글 페이지** | **글 목록 페이지** | **MyPage** (작성글) | **MemberPage** | **삭제한 게시글 목록** | **삭제한 게시글 페이지** |
+| :--: | :--: | :--: | :--: | :--: | :--: | :--: |
 | id <small>게시글 아이디<small> |  |  | O | O |  |  |
 | authorId <small>작성자 아이디<small> | O (name) | O (name) | (댓글단 글, 좋아요한 글) |  |  |  |
 | createdAt <small>작성 시각<small> | O | O | O | O |  | O |
@@ -56,8 +67,8 @@ Types.tsx에 필요해보이는 type을 정리해두긴 했는데, 뭉치(?)가 
 | commentCount <small>댓글 수<small> | (O) | O (count) | O (count) | O (count) |  |  |
 
 
-| ==**Comment**== | **게시글 페이지** | **MyPage (작성댓글)** | **MemberPage** | (**글 목록 페이지**) |
-| ---- | ---- | ---- | ---- | ---- |
+| ***Comment*** | **게시글 페이지** | **MyPage** (작성댓글) | **MemberPage** | (**글 목록 페이지**) |
+| :--: | :--: | :--: | :--: | :--: |
 | id <small>댓글 아이디<small> |  |  | O |  |
 | authorId <small>작성자 아이디<small> | O (name) |  |  |  |
 | createdAt <small>작성 시각<small> | O | O | O |  |
@@ -69,11 +80,13 @@ Types.tsx에 필요해보이는 type을 정리해두긴 했는데, 뭉치(?)가 
 백엔드에서 상위댓글인지 하위댓글인지에 대한 정보를 어떻게 전달해줄지 몰라서 일단은 priorComment(상위 댓글 아이디)를 새로 넣었습니다. 그렇게 하면 reComment를 굳이 함께 전달해주지 않아도 괜찮을 것 같아요.
 
 
-==**Board(게시판)**== , ==**Cafe(카페)**== 의 경우 작성된 api 명세표대로 하면 될 듯하고, cafe의 경우도 추후 조금 달라질 수 있지만 미리 구현해두어서 적지 않았습니다.
+***Board(게시판)*** , ***Cafe(카페)*** 의 경우 작성된 api 명세표대로 하면 될 듯하고, cafe의 경우도 추후 조금 달라질 수 있지만 미리 구현해두어서 적지 않았습니다.
 
 
 
-**요약**:
+
+##### **요약**
+
 Context API는 로그인, 회원가입에 사용하고, 전체적으로 자주 참고하는 상태(state)가 생기면 그에 맞춰 사용하면 될 것 같습니다. (ex - 모달 버튼의 클릭 여부 등...)
 
 Types.tsx에 적힌 것보단 여기에 적힌 내용을 참고하시면 됩니다.
